added player scan fix, made transmog fix less strict for WotLK gear
will ignore tooltips of npcs (if a player has the same name as an npc, the tooltip of that npc used to show the players GS).
will ignore players with the "UNKNOWN" temp name after login when not all players are fully loaded yet.

will ignore players' gear for the following reasons:
- Item ID is below 35560 (a rough cut-off for when item ids for wotlk start)
- Item Quality is poor or common
- Item Level is below 130 (with the exception of heirlooms, which are item lvl 1)

This will catch all incorrect Player Scans that contain pre-WotLK gear or WotLK gear, that is clearly too low.
This will NOT catch transmog of WotLK gear. (if someone transmogs their T10 to look like T7, this will not be caught)

This will probably catch MOST, but not all incorrcect Player Scans (as well as received Scans).